Break ref cycle between API::HTTPCookieStore and WebKit::WebsiteDataStore https://bugs.webkit.org/show_bug.cgi?id=226992
Reviewed by Chris Dumez.

Source/WebKit:

Covered by an API test.

* UIProcess/API/APIHTTPCookieStore.cpp:
(API::HTTPCookieStore::HTTPCookieStore):
(API::HTTPCookieStore::filterAppBoundCookies):
(API::HTTPCookieStore::cookies):
(API::HTTPCookieStore::cookiesForURL):
(API::HTTPCookieStore::setCookies):
(API::HTTPCookieStore::deleteCookie):
(API::HTTPCookieStore::deleteAllCookies):
(API::HTTPCookieStore::setHTTPCookieAcceptPolicy):
(API::HTTPCookieStore::flushCookies):
(API::HTTPCookieStore::registerObserver):
(API::HTTPCookieStore::unregisterObserver):
* UIProcess/API/APIHTTPCookieStore.h:
* UIProcess/Cocoa/SOAuthorization/SOAuthorizationSession.mm:
(WebKit::SOAuthorizationSession::complete):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/WKWebsiteDatastore.mm:
(TestWebKitAPI::TEST):



Canonical link: https://commits.webkit.org/238994@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279074 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/UIProcess/API/APIHTTPCookieStore.cpp
@@ -44,7 +44,7 @@ using namespace WebKit;
 namespace API {
 
 HTTPCookieStore::HTTPCookieStore(WebKit::WebsiteDataStore& websiteDataStore)
-    : m_owningDataStore(websiteDataStore)
+    : m_owningDataStore(makeWeakPtr(websiteDataStore))
 {
 }
 
@@ -55,45 +55,52 @@ HTTPCookieStore::~HTTPCookieStore()
     ASSERT(!m_cookieManagerProxyObserver);
 }
 
-void HTTPCookieStore::filterAppBoundCookies(const Vector<WebCore::Cookie>& cookies, CompletionHandler<void(Vector<WebCore::Cookie>&&)>&& completionHandler)
+void HTTPCookieStore::filterAppBoundCookies(Vector<WebCore::Cookie>&& cookies, CompletionHandler<void(Vector<WebCore::Cookie>&&)>&& completionHandler)
 {
-    Vector<WebCore::Cookie> appBoundCookies;
 #if ENABLE(APP_BOUND_DOMAINS)
-    m_owningDataStore->getAppBoundDomains([cookies, appBoundCookies = WTFMove(appBoundCookies), completionHandler = WTFMove(completionHandler)] (auto& domains) mutable {
+    if (!m_owningDataStore)
+        return completionHandler({ });
+    m_owningDataStore->getAppBoundDomains([cookies = WTFMove(cookies), completionHandler = WTFMove(completionHandler)] (auto& domains) mutable {
+        Vector<WebCore::Cookie> appBoundCookies;
         if (!domains.isEmpty() && !isFullWebBrowser()) {
-            for (auto& cookie : cookies) {
+            for (auto& cookie : WTFMove(cookies)) {
                 if (domains.contains(WebCore::RegistrableDomain::uncheckedCreateFromHost(cookie.domain)))
-                    appBoundCookies.append(cookie);
+                    appBoundCookies.append(WTFMove(cookie));
             }
         } else
-            appBoundCookies = cookies;
+            appBoundCookies = WTFMove(cookies);
         completionHandler(WTFMove(appBoundCookies));
     });
 #else
-    appBoundCookies = cookies;
-    completionHandler(WTFMove(appBoundCookies));
+    completionHandler(WTFMove(cookies));
 #endif
 }
 
 void HTTPCookieStore::cookies(CompletionHandler<void(const Vector<WebCore::Cookie>&)>&& completionHandler)
 {
+    if (!m_owningDataStore)
+        return completionHandler({ });
     auto& cookieManager = m_owningDataStore->networkProcess().cookieManager();
-    cookieManager.getAllCookies(m_owningDataStore->sessionID(), [this, protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)] (const Vector<WebCore::Cookie>& cookies) mutable {
-        filterAppBoundCookies(cookies, WTFMove(completionHandler));
+    cookieManager.getAllCookies(m_owningDataStore->sessionID(), [this, protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)] (Vector<WebCore::Cookie>&& cookies) mutable {
+        filterAppBoundCookies(WTFMove(cookies), WTFMove(completionHandler));
     });
 }
 
 void HTTPCookieStore::cookiesForURL(WTF::URL&& url, CompletionHandler<void(Vector<WebCore::Cookie>&&)>&& completionHandler)
 {
+    if (!m_owningDataStore)
+        return completionHandler({ });
     auto& cookieManager = m_owningDataStore->networkProcess().cookieManager();
     cookieManager.getCookies(m_owningDataStore->sessionID(), url, [this, protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)] (Vector<WebCore::Cookie>&& cookies) mutable {
-        filterAppBoundCookies(cookies, WTFMove(completionHandler));
+        filterAppBoundCookies(WTFMove(cookies), WTFMove(completionHandler));
     });
 }
 
-void HTTPCookieStore::setCookies(const Vector<WebCore::Cookie>& cookies, CompletionHandler<void()>&& completionHandler)
+void HTTPCookieStore::setCookies(Vector<WebCore::Cookie>&& cookies, CompletionHandler<void()>&& completionHandler)
 {
-    filterAppBoundCookies(cookies, [this, protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)] (auto&& appBoundCookies) mutable {
+    filterAppBoundCookies(WTFMove(cookies), [this, protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)] (auto&& appBoundCookies) mutable {
+        if (!m_owningDataStore)
+            return;
         auto& cookieManager = m_owningDataStore->networkProcess().cookieManager();
         cookieManager.setCookies(m_owningDataStore->sessionID(), appBoundCookies, WTFMove(completionHandler));
     });
@@ -101,12 +108,16 @@ void HTTPCookieStore::setCookies(const Vector<WebCore::Cookie>& cookies, Complet
 
 void HTTPCookieStore::deleteCookie(const WebCore::Cookie& cookie, CompletionHandler<void()>&& completionHandler)
 {
+    if (!m_owningDataStore)
+        return completionHandler();
     auto& cookieManager = m_owningDataStore->networkProcess().cookieManager();
     cookieManager.deleteCookie(m_owningDataStore->sessionID(), cookie, WTFMove(completionHandler));
 }
 
 void HTTPCookieStore::deleteAllCookies(CompletionHandler<void()>&& completionHandler)
 {
+    if (!m_owningDataStore)
+        return completionHandler();
     auto& cookieManager = m_owningDataStore->networkProcess().cookieManager();
     cookieManager.deleteAllCookies(m_owningDataStore->sessionID());
     // FIXME: The CompletionHandler should be passed to WebCookieManagerProxy::deleteAllCookies.
@@ -115,12 +126,16 @@ void HTTPCookieStore::deleteAllCookies(CompletionHandler<void()>&& completionHan
 
 void HTTPCookieStore::setHTTPCookieAcceptPolicy(WebCore::HTTPCookieAcceptPolicy policy, CompletionHandler<void()>&& completionHandler)
 {
+    if (!m_owningDataStore)
+        return completionHandler();
     auto& cookieManager = m_owningDataStore->networkProcess().cookieManager();
     cookieManager.setHTTPCookieAcceptPolicy(m_owningDataStore->sessionID(), policy, WTFMove(completionHandler));
 }
 
 void HTTPCookieStore::flushCookies(CompletionHandler<void()>&& completionHandler)
 {
+    if (!m_owningDataStore)
+        return completionHandler();
     m_owningDataStore->flushCookies(WTFMove(completionHandler));
 }
 
@@ -145,7 +160,7 @@ void HTTPCookieStore::registerObserver(Observer& observer)
 {
     m_observers.add(&observer);
 
-    if (m_cookieManagerProxyObserver)
+    if (m_cookieManagerProxyObserver || !m_owningDataStore)
         return;
 
     ASSERT(!m_observedCookieManagerProxy);
@@ -163,7 +178,7 @@ void HTTPCookieStore::unregisterObserver(Observer& observer)
     if (!m_observers.computesEmpty())
         return;
 
-    if (m_observedCookieManagerProxy)
+    if (m_observedCookieManagerProxy && m_owningDataStore)
         m_observedCookieManagerProxy->unregisterObserver(m_owningDataStore->sessionID(), *m_cookieManagerProxyObserver);
 
     m_observedCookieManagerProxy = nullptr;
